SetUserID Dispatch
