+++ src/client/styles/App.scss
@@ -5,17 +5,19 @@ h1 {
 }
 
 .login-wrapper {
-    margin: auto;
     display: flex;
     justify-content: center;
     align-items: center;
+    height: 50%;
+    width: min-content;
 }
 
 .signup-wrapper {
-    margin: auto;
     display: flex;
     justify-content: center;
     align-items: center;
+    height: 50%;
+    width: 80%;
 }
 
 .profile-wrapper {
@@ -36,7 +38,7 @@ h1 {
 @media screen and (max-width: 890px) {
     .form {
         @include form;
-        min-width: 80%;
+        min-width: 40%;
     }
 }
 @media screen and (min-width: 891px) {
@@ -62,20 +64,31 @@ h1 {
     width: 100%;
 }
 
-.form > h1 {
+.headersContainer {
+    width: 110%;
+    hr {
+        padding: 0;
+        border: 1px solid rgb(0, 0, 0);
+        box-shadow: 0px 0px 5px 0px #212121;
+    }
+}
+
+.headersContainer > .loginHeader {
     font-family: Montserrat, sans-serif;
     text-transform: uppercase;
     font-weight: 400;
     margin: 0px 0px 1rem 0px;
     padding: 5px;
     margin-left: auto;
-    text-decoration: underline;
     margin-right: auto;
     border-radius: 5px;
-    color: #ffffff;
+    background: linear-gradient(90deg, rgba(0, 217, 233, 0.865) 0%, rgba(48, 45, 255, 0.51) 100%);
+    background-clip: text;
+    -webkit-background-clip: text;
+    -webkit-text-fill-color: transparent;
+    padding: 0px;
     text-align: center;
-    background-color: #4040407c;
-    box-shadow: 0px 0px 5px 0px #212121;
+    box-shadow: -4px 4px 1px 1px #212121;
     width: fit-content;
     height: fit-content;
 }
@@ -124,4 +137,10 @@ h1 {
     justify-content: center;
     align-items: center;
     margin: 5px;
+}
+
+.outerContainer {
+    @include flex-vert;
+    width: 100%;
+    height: 100%;
 }
