+++ src/client/styles/Main.scss
@@ -38,10 +38,22 @@ $gray: #92959E;
     }
 }
 
+@keyframes move-down-into-middle {
+    0% {
+        transform: translateY(-100vh);
+    }
+    100% {
+        transform: translateY(0);
+    }
+}
+
 .outerContainerMain {
     width: 100%;
     height: 100%;
-    
+}
+
+.outerContainerMain.animated {
+    animation: move-down-into-middle 1s ease-in-out;
 }
