+++ src/client/styles/Profile.scss
@@ -15,6 +15,7 @@ p {
     background-color: #FFFFFF;
     border-radius: 5px;
     box-shadow: 0px 10px 20px -10px rgba(0, 0, 0, 0.75);
+    border: 2px solid #000000;
     color: #B3B8CD;
     padding-top: 30px;
     padding-bottom: 30px;
@@ -23,7 +24,6 @@ p {
     min-width: 350px;
     text-align: center;
     margin: 0 auto;
-    margin-top: 10%;
     background-image: url("https://img.rawpixel.com/s3fs-private/rawpixel_images/website_content/v1003-71c-x.jpg?w=800&dpr=1&fit=default&crop=default&q=65&vib=3&con=3&usm=15&bg=F4F4F3&ixlib=js-2.2.1&s=4e6ef20121f4996cef83ee6eb410c841");
     background-size: cover;
     background-repeat: no-repeat;
